added new cropping gui
